// vendor-src/folia/src/types/navidrome.ts
export interface SubsonicSong {
    id: string;
    parent?: string;
    isDir: boolean;
    title: string;
    album: string;
    artist: string;
    track?: number;
    year?: number;
    genre?: string;
    coverArt?: string;
    size: number;
    contentType: string;
    suffix: string;
    duration: number; // seconds
    bitRate?: number;
    path: string;
    playCount?: number;
    discNumber?: number;
    created: string;
    albumId: string;
    artistId: string;
    type: 'music' | 'podcast' | 'audiobook';
    isVideo: boolean;
    starred?: string;
    replayGain?: ReplayGainInfo;
}